Quynh

+++ src/main/java/ptit/repository/AccountDAO.java
@@ -1,6 +1,6 @@
 package ptit.repository;
 
-import java.math.BigInteger;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
 import ptit.entity.Account;
 
 @Repository
-public interface AccountDAO extends JpaRepository<Account, BigInteger> {
+public interface AccountDAO extends JpaRepository<Account, Long> {
 
 	@Query("from Account as u where u.email = :email")
 	Optional<Account> findByEmail(@Param("email") String email);
